build: move top level dependencies into gdal package
