+++ scripts/generate_hillshade.py
@@ -15,10 +15,10 @@ from topo_imagery_common.log.time_helper import time_in_ms
 from topo_imagery_gdal.gdal.gdal_commands import get_gdal_command, get_hillshade_command
 from topo_imagery_gdal.gdal.gdal_helper import run_gdal
 from topo_imagery_gdal.gdal.gdal_presets import CompressionPreset, HillshadePreset
+from topo_imagery_gdal.standardising import create_vrt
 
 from scripts.json_codec import dict_to_json_bytes
 from scripts.stac.imagery.create_stac import create_item
-from scripts.standardising import create_vrt
 
 
 def get_args_parser() -> CommonArgumentParser:
