+++ scripts/standardise_validate.py
@@ -17,11 +17,11 @@ from topo_imagery_common.datetimes import RFC_3339_DATETIME_FORMAT, format_rfc_3
 from topo_imagery_common.files.files_helper import SUFFIX_JSON, ContentType
 from topo_imagery_common.files.fs import exists, write
 from topo_imagery_gdal.gdal.gdal_helper import get_srs, get_vfs_path
+from topo_imagery_gdal.standardising import StandardisingConfig, run_standardising
 from topo_imagery_gdal.tiff.file_tiff import FileTiff
 
 from scripts.json_codec import dict_to_json_bytes
 from scripts.stac.imagery.create_stac import create_item
-from scripts.standardising import StandardisingConfig, run_standardising
 
 
 def get_args_parser() -> ArgumentParser:
